Renamed namespaces and folders from "Windows" to "Forms"
