+++ Aquality.WinAppDriver/tests/Aquality.WinAppDriver.Tests/Actions/MouseActionsTests.cs
@@ -1,7 +1,7 @@
 ﻿using Aquality.WinAppDriver.Actions;
 using Aquality.WinAppDriver.Applications;
 using Aquality.WinAppDriver.Elements.Interfaces;
-using Aquality.WinAppDriver.Tests.Windows;
+using Aquality.WinAppDriver.Tests.Forms;
 using NUnit.Framework;
 
 namespace Aquality.WinAppDriver.Tests.Actions
