+++ Aquality.WinAppDriver/tests/Aquality.WinAppDriver.Tests/Elements/ElementFactoryTests.cs
@@ -1,7 +1,7 @@
 ﻿using Aquality.WinAppDriver.Applications;
 using Aquality.WinAppDriver.Elements;
 using Aquality.WinAppDriver.Elements.Interfaces;
-using Aquality.WinAppDriver.Tests.Windows;
+using Aquality.WinAppDriver.Tests.Forms;
 using NUnit.Framework;
 using OpenQA.Selenium;
 using System;
